refactor(naming): name the transaction-group module `group`
`algonaut_transaction` already carries the noun, so the module need not
repeat it (Rule 1) — `transaction::group::TransactionGroup` reads without
the threefold repetition of `transaction::transaction_group::TransactionGroup`,
and matches the existing `atomic::group` submodule. The public type stays
`TransactionGroup`. File `transaction_group.rs` → `group.rs`; ADRs and
CHANGELOG updated to match.
