+++ CHANGELOG.md
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Breaking:** the `algonaut::atomic_transaction_composer` module is renamed to `algonaut::atomic`. Update imports: `use algonaut::atomic_transaction_composer::{…}` → `use algonaut::atomic::{…}`. No type names, signatures, or behaviour change — the directory was named after the `AtomicTransactionComposer` type the typestate refactor deleted. The module is also reorganized internally into `group`, `method_call`, `encode`, `outcome`, and `signing` submodules (private; the flat public re-exports are unchanged). See `docs/adr/atomic-module-layout.md`
 - **Breaking:** the closed `algonaut::atomic_transaction_composer::transaction_signer::TransactionSigner` enum is replaced by the open `Signer` trait. `TransactionWithSigner.signer` is now `Option<Arc<dyn Signer>>` (`None` mirrors the old `TransactionSigner::Empty` simulate slot — the composer fills it with an all-zero placeholder signature). `AddMethodCallParams.signer` is now `Arc<dyn Signer>`. Migrate constructions: `TransactionSigner::BasicAccount(acc)` → `Arc::new(acc) as Arc<dyn Signer>`; `TransactionSigner::ContractAccount(ca)` → `Arc::new(ca)`; `TransactionSigner::MultisigAccount { address, accounts }` → `Arc::new(MultisigSigner { address, accounts })`; `TransactionSigner::Empty` → drop, use `TransactionWithSigner::unsigned(tx)` or `signer: None`
 - **Breaking:** the `MethodCall` builder takes the method and its arguments together via `.invoke(...)` instead of the separate `.args(...)` setter, and `MethodCall::builder` no longer takes the method positionally. Migrate `MethodCall::builder(app_id, method, sender, signer).args([2u64, 3u64])` → `MethodCall::builder(app_id, sender, signer).invoke(abi_call!("add(uint64,uint64)uint64", 2u64, 3u64))` for compile-time-literal signatures, or `.invoke(Invocation::new(method, [2u64, 3u64]))` when the method comes from `AbiMethod::from_signature` at run time. See `docs/adr/abi-method-signature-macro.md`
-- **Breaking:** transaction naming is normalized across the handwritten API — the `Transaction` noun is dropped where the receiver or argument type already carries it, and otherwise spelled out in full (never `tx` / `txn` / `txid` in a public name). Type renames: `TxId` → `TransactionId`, `TxGroup` → `TransactionGroup` (module `algonaut_transaction::tx_group` → `transaction_group`), `TxnHeader` → `TransactionHeader`. `Algod` methods: `send_txn` → `send`, `send_txns` → `send_transactions`, `send_txn_async` → `send_async`, `send_txns_async` → `send_transactions_async`, `send_raw_txn` → `send_raw`, `send_raw_txn_async` → `send_raw_async`, `pending_txn` → `pending_transaction`, `pending_txns` → `pending_transactions`, `address_pending_txns` → `address_pending_transactions`, `block_txids` → `block_transaction_ids`, `txn_proof` → `transaction_proof`, `txn_group_state_delta` → `transaction_group_state_delta`, `txn_group_state_deltas_for_round` → `transaction_group_state_deltas_for_round`, `submit_txns` → `submit_transactions`, `simulate_txns` → `simulate`, `txn_params` → `suggested_params`. `Account::sign_transaction` and `Kmd::sign_transaction` → `sign`. The transaction-id concept is spelled `transaction_id` everywhere — `txid` / `tx_id` parameters and fields retire, including `PendingSubmission::tx_id()` → `transaction_id()` and the `tx_id` / `tx_ids` fields on the atomic `ExecuteOutcome` / `SimulateOutcome` / `AbiMethodResult`. Amends the `TxId` / `TxGroup` names introduced by the identifier-newtypes ADR. Rename-only: no behaviour or wire-format change. See `docs/adr/transaction-naming-convention.md`
+- **Breaking:** transaction naming is normalized across the handwritten API — the `Transaction` noun is dropped where the receiver or argument type already carries it, and otherwise spelled out in full (never `tx` / `txn` / `txid` in a public name). Type renames: `TxId` → `TransactionId`, `TxGroup` → `TransactionGroup` (module `algonaut_transaction::tx_group` → `group`), `TxnHeader` → `TransactionHeader`. `Algod` methods: `send_txn` → `send`, `send_txns` → `send_transactions`, `send_txn_async` → `send_async`, `send_txns_async` → `send_transactions_async`, `send_raw_txn` → `send_raw`, `send_raw_txn_async` → `send_raw_async`, `pending_txn` → `pending_transaction`, `pending_txns` → `pending_transactions`, `address_pending_txns` → `address_pending_transactions`, `block_txids` → `block_transaction_ids`, `txn_proof` → `transaction_proof`, `txn_group_state_delta` → `transaction_group_state_delta`, `txn_group_state_deltas_for_round` → `transaction_group_state_deltas_for_round`, `submit_txns` → `submit_transactions`, `simulate_txns` → `simulate`, `txn_params` → `suggested_params`. `Account::sign_transaction` and `Kmd::sign_transaction` → `sign`. The transaction-id concept is spelled `transaction_id` everywhere — `txid` / `tx_id` parameters and fields retire, including `PendingSubmission::tx_id()` → `transaction_id()` and the `tx_id` / `tx_ids` fields on the atomic `ExecuteOutcome` / `SimulateOutcome` / `AbiMethodResult`. Amends the `TxId` / `TxGroup` names introduced by the identifier-newtypes ADR. Rename-only: no behaviour or wire-format change. See `docs/adr/transaction-naming-convention.md`
 
 ## [0.7.0] - 2026-05-20
